The problem with the communication was fixed
Same STATUS messages of the ACS was being lost. So, i added a larger value for the timeout of the read_tcpip_function. Also, I modified the state machine of the GEI. Some "DONE" messages of the ACS were being accumulated, creating a delay in the indicators of the interface